Refactored instance and admin components
